+++ tests/test_tdjson.py
@@ -7,13 +7,13 @@ class Test_get_tdjson_lib_path(object):
         mocked_resource = mocker.Mock()
 
         with mocker.mock_module.patch('telegram.tdjson.platform.system', mocked_system):
-            with mocker.mock_module.patch('telegram.tdjson.pkg_resources.resource_filename',
-                                          mocked_resource):
+            with mocker.mock_module.patch(
+                'telegram.tdjson.pkg_resources.resource_filename', mocked_resource
+            ):
                 _get_tdjson_lib_path()
 
         mocked_resource.assert_called_once_with(
-            'telegram',
-            'lib/darwin/libtdjson.dylib',
+            'telegram', 'lib/darwin/libtdjson.dylib'
         )
 
     def test_for_linux(self, mocker):
@@ -21,39 +21,33 @@ class Test_get_tdjson_lib_path(object):
         mocked_resource = mocker.Mock(return_value='/tmp/')
 
         with mocker.mock_module.patch('telegram.tdjson.platform.system', mocked_system):
-            with mocker.mock_module.patch('telegram.tdjson.pkg_resources.resource_filename',
-                                          mocked_resource):
+            with mocker.mock_module.patch(
+                'telegram.tdjson.pkg_resources.resource_filename', mocked_resource
+            ):
                 _get_tdjson_lib_path()
 
-        mocked_resource.assert_called_once_with(
-            'telegram',
-            'lib/linux/libtdjson.so',
-        )
+        mocked_resource.assert_called_once_with('telegram', 'lib/linux/libtdjson.so')
 
     def test_for_windows(self, mocker):
         mocked_system = mocker.Mock(return_value='Windows')
         mocked_resource = mocker.Mock(return_value='/tmp/')
 
         with mocker.mock_module.patch('telegram.tdjson.platform.system', mocked_system):
-            with mocker.mock_module.patch('telegram.tdjson.pkg_resources.resource_filename',
-                                          mocked_resource):
+            with mocker.mock_module.patch(
+                'telegram.tdjson.pkg_resources.resource_filename', mocked_resource
+            ):
                 _get_tdjson_lib_path()
 
-        mocked_resource.assert_called_once_with(
-            'telegram',
-            'lib/linux/libtdjson.so',
-        )
+        mocked_resource.assert_called_once_with('telegram', 'lib/linux/libtdjson.so')
 
     def test_unknown(self, mocker):
         mocked_system = mocker.Mock(return_value='Unknown')
         mocked_resource = mocker.Mock(return_value='/tmp/')
 
         with mocker.mock_module.patch('telegram.tdjson.platform.system', mocked_system):
-            with mocker.mock_module.patch('telegram.tdjson.pkg_resources.resource_filename',
-                                          mocked_resource):
+            with mocker.mock_module.patch(
+                'telegram.tdjson.pkg_resources.resource_filename', mocked_resource
+            ):
                 _get_tdjson_lib_path()
 
-        mocked_resource.assert_called_once_with(
-            'telegram',
-            'lib/linux/libtdjson.so',
-        )
+        mocked_resource.assert_called_once_with('telegram', 'lib/linux/libtdjson.so')