+++ telegram/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '0.8.0'
+__version__ = '0.9.0'
 
 VERSION = __version__