+++ docs/source/changelog.rst
@@ -4,7 +4,8 @@ Changelog
 
 [unreleased]
 
-- Fixed problem with randomly raised "Database encryption key is needed" errors during login process.
+- Fixed problem with randomly raised "Database encryption key is needed" errors during login process. (#12)
+- Fixed `stop` method execution. (#8)
 
 [0.8.0] - 2019-03-17